Added: deleting vegetable by name, deleting vegetable by color

+++ src/Main/CaloriesCalculating.java
@@ -2,7 +2,7 @@ import java.util.ArrayList;
 
 public class CaloriesCalculating {
 
-    public static double sumOfCalories(ArrayList<Vegetable> vegetableArrayList) {
+    public static int sumOfCalories(ArrayList<Vegetable> vegetableArrayList) {
         int sum = 0;
         for (Vegetable v : vegetableArrayList) {
             sum += v.getCalories();
